Want to experiment with content from class
Learned about Hinge loss and linear classification the other day, would like to try it out

+++ lib/mnist_csv.c
@@ -15,7 +15,7 @@ int get_next_data(struct MnistCSV* csv) {
 		char c = fgetc(csv->file);
 		if (c == ',' || (c == '\n' && charCount != 0)) {
 			digitString[charCount] = '\0';
-			csv->buffer[index] = atoi(digitString);
+			csv->buffer[index] = atof(digitString);
 			charCount = 0;
 			index++;
 		} else if (c != '\n') {
@@ -28,9 +28,9 @@ int get_next_data(struct MnistCSV* csv) {
 }
 
 void visualize_digit_data(struct MnistCSV* csv) {
-	int* digit = csv->buffer;
+	float* digit = csv->buffer;
 	printf("============================\n");
-	printf("Data for digit %d:\n", digit[0]);
+	printf("Data for digit %f:\n", digit[0]);
 	for (int i = 0; i < 28; i++) {
 		for (int j = 0; j < 28; j++) {
 			if (digit[i * 28 + j + 1] < 80) {
