+++ lib/mnist_csv.h
@@ -5,10 +5,10 @@
 // Buffer should have space for 785 ints
 struct MnistCSV {
 	FILE* file;
-	int* buffer;
+	float* buffer;
 };
 
-// Populates the integer buffer array where the 1st entry is the numerical value and the remaining 784 are pixel values 0-255. Returns 1 if CSV file is empty, 0 otherwise
+// Populates the integer buffer array with a single row of data where the 1st entry is the numerical value and the remaining 784 are pixel values 0-255. Returns 1 if CSV file is empty, 0 otherwise
 int get_next_data(struct MnistCSV* csv);
 void visualize_digit_data(struct MnistCSV* csv);
